requirements.txt
streamlit
gitpython
chromadb
sentence-transformers
openai
beautifulsoup4
